requirements.txt
Flask==2.1.2
tensorflow-cpu
gunicorn
numpy==1.26.0
stanza==1.9.2
Sastrawi==1.0.1
gensim==4.3.3
pandas==2.2.3
scikit-learn==1.2.0
werkzeug==2.0.3

